Adds recipe creation endpoint
Introduces a new endpoint for creating recipes, including handling of recipe ingredients.

Migrates recipe and recipe ingredient IDs to use integers instead of bigints.

+++ backend-rv/drizzle/0000_dark_loa.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "recipe_ingredients" (
 	"id" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name "recipe_ingredients_id_seq" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),
-	"recipe_id" bigint NOT NULL,
+	"recipe_id" integer NOT NULL,
 	"name" text NOT NULL,
 	"quantity" text NOT NULL,
 	"unit" text NOT NULL,
@@ -11,7 +11,7 @@ CREATE TABLE "recipe_ingredients" (
 --> statement-breakpoint
 CREATE TABLE "recipes" (
 	"id" integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY (sequence name "recipes_id_seq" INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START WITH 1 CACHE 1),
-	"user_id" bigint NOT NULL,
+	"user_id" integer NOT NULL,
 	"name" text NOT NULL,
 	"instructions" text[] DEFAULT '{}'::text[] NOT NULL
 );
